+++ matchmaker-cli/src/config.rs
@@ -52,11 +52,6 @@ pub struct Config {
     #[serde(default)]
     #[partial(attr, alias = "e")]
     pub exit: ExitConfig,
-
-    // semantic aliases
-    #[serde(default)]
-    #[partial(skip)]
-    pub aliases: HashMap<String, Trigger>,
 }
 
 // -----------------------